Consistent formatValue
